Fix tcp_segmentation_offload
Due to nature of loading yaml in Python, TCP segmentation offload value
in the config.yaml is converted to True or False in the inventory.yml
and downstream in the /etc/network/interfaces file on the operating systems.

The fix is to set the off value as a string to prevent the automatic
string to boolean conversion.

Change-Id: Ie915e1430fe20c16429d3cdedd2659baf911094b