fix: Set MTU property for all interface types
The first issue resolved by this commit pertains to the MTU setting in
the interfaces_ubuntu.j2 template. The directive was placed within a
loop defining bonded interfaces only. The MTU related code is now
defined such that it is applied for all interface types.

The second issue relates to the way Ubuntu handles MTU sizes on VLANs
and bridges. If a VLAN and/or bridge is brought up before it's
interface depencies have been met with respect to MTU size, the VLAN
and/or bridge will ignore it's specified MTU size and use the default
value.

Change-Id: I483cc788122246364383d6d43f88e60181c98256