feat: Add parallel power on/off sequencing
Power on/off was performed synchronously and sequentially. Each
system had to complete before the action was launched on the next
system.

Now the task is launched in parallel and the power state is polled until
all the systems are in the desired state or the timeout period has been
reached.

The playbooks have also been set to no longer ignore errors.

Change-Id: Ida8fbf3f556842ed7c5a807c4c15c8a29e6e763a