feat: Generate 'known_hosts' with SSH host keys
Generate a 'known_hosts' file containing SSH host keys for the
deployment container and all client nodes. This file is stored locally
in the 'cluster-genesis/playbooks' directory. Its content is appended to
the '~/.ssh/known_hosts' files of the deployer user locally and of root
user on all client nodes.

Change-Id: I758f4d101d273df9f1c9477fc2922c50f436cf87