feat: Multiple bootstrap_cmds for multiple hosts
software bootstrapping in genesis was previously limited to a single set
of hosts, and a single set of commands. This commit will eliminate the
software-bootstrap-hosts and software-bootstrap-cmd keys, and replace
them with a software-bootstrap dictionary. The keys for this dictionary
will be the host(s) to target and the value will be the command to run.
The host can be a group, or a subset of groups, i.e. all, or controller,
or controller[0] etc.

Change-Id: I5c2b7d2a380106287c5aaec122d298d159602486