fix: Configure bonded interfaces in Ubuntu
When bonds are defined in one or more of a node's networks the 'bonding'
kernel module is added to '/etc/modules' so that it will automatically
load upon boot.

Bonding in Ubuntu requires the software package 'ifenslave'. The Ubuntu
preseed files have been updated to include this package during the
initial OS installation.

MAC address assignment to bond interfaces (and by extension any bridges
or vlans assigned to the bond) can be inconsistent. Directives are now
included in the '/etc/network/interfaces' file to set the bond MAC (and
all derivative interface MACs) to match the 'bond-primary' interface as
designated in the 'config.yml' network bond definition (or if
'bond-primary' doesn't exist it will use the first interface from the
'bond-interfaces' list).

Change-Id: I70b8908429042b8e08393e48be24e02cb324bc92