fix: Set mock 'inventory.yml' access mode to 0644
Previously the 'inventory.yml' output file mode was set to match that of
the 'config.yml' input file. If the latter was not writable the script
would fail.

The 'config.yml' input file only needs to be readable. The
'inventory.yml' output file will always have access mode set to '0644'.

Change-Id: I95d135e7df78042b76c4f99323f18b0c068dd335