feat: Enable two additional data net interfaces
Up to four data network interfaces can be assigned to each host.
Interfaces 'eth12' and 'eth13' have been enabled in addition to
the existing 'eth10' and 'eth11'. If the interfaces are configured
without any bonding then some, none or all may be configured. When
configuring bonds in conjunction with LAG/MLAG then eth10/eth11 and
eth12/13 must be paired together.

Change-Id: Iad2f0ed348fec5716f3df8d13904957dc697326e