fix: Ensure deployer log file is owned by user
The local log file on the deployer ('cluster-genesis/log.txt') was being
created by a script run with 'sudo'. This resulted in the log file being
owned by the root user. Any subsequent scripts which utilized the logger
but weren't being run through 'sudo' would fail due to improper
permissions while attempting to append to the log file. Now a blank log
file, owned by the user running the playbooks, is created at the
beginning of the first play. Subsequent logs are appended to this file
without changing ownership or mode.

Change-Id: I152f63728b0e4af06bb890253a06f2ed380b465e