feat: Add support for non-tagged mgmt interface
The switch will be accessed through its non-tagged interface in order to
configure the management network.

The network device attached to the switch can be specified in the
'config.yml' file. If it is not specified the network devices will be
traversed until the switch can be accessed through a ping.

No configuration takes place if the switch can be accessed through the
management bridge.

Change-Id: I718754621578e6fe76c6d131c48c436ee616aad2