feat: Integrate ppc64el introspection support
Introspection consists of loading a lightweight in-memory OS (linux
buildroot) on all client nodes prior to OS installation to disk. This
feature can be enabled by setting the 'introspection-enabled' key in
'config.yml' to a boolean value. Initially it is only supported on
clusters with all ppc64el deployer and client nodes.

Change-Id: I1539110ccf797b08ab6acbe3e7d6351ed49cf841