feat: Track OS disk path(s) per node
The 'os-disk' key can optionally be included within the 'config.yml'
'node-templates' dictionary. The value is set to the path of the disk to
be used for OS installation (or in the case of software RAID a list of
paths). The scope of this key/value applies to all members of the
respective node template. This commit changes the scope to individual
nodes. This allows the value to be updated during genesis (most likely
during "introspection") to values specific to each node.

Change-Id: Ieb978defe3b5e1e70ad66d9c38e74371fd2e91c3