feat: Allow deployment without data switch access
Data switches can be set to "passive" mode to allow deployment without
supplying login credentials to the switch management interfaces. This
mode requires the user to manually write switch MAC address tables to
file and to configure the data switch in accordance with the defined
networks. The client networks will still be configured by Cluster
Genesis.

To configure passive switches simply omit 'userid-data-switch' from
'config.yml'. The 'ipaddr-data-switch' dictionary still needs to be
defined in order to be used as a switch identifier. In place of IP
addresses anything may be used as long as each switch has a unique
value.

Change-Id: I0bfc305bbedb404c55eb7f86021e4d8e84a1f7ac