feat: Add 'tear-down' switch host key removal
The 'tear-down' script will optionally remove management and/or data
switch host keys.

This is useful when, for example, data switches are configured for MLAG
and their host keys are both set to match the primary switch. If a user
saves the resulting host keys to their ~/.ssh/known_hosts they would
presumably also want to remove those keys when the MLAG configuration
is removed.

Change-Id: Id5f3644cdfb7f449c3f1566a0b6c236ce3650577