feat: Allow configuring the deployer as a gateway
If the 'deployer-gateway' key in 'config.yml' is set to 'true' then the
deployer will be configured as the default gateway for all client nodes.
Configuration includes adding a 'MASQUERADE' rule to the deployer's
'iptables' NAT chain and setting the 'dnsmasq' DHCP service to serve the
deployer's client management bridge address as the default gateway. Note
that setting the 'gateway' explicitly on any of the data networks will
override this.

Change-Id: Ifabb5b390f04d7792814411b5a016a97bb08da24