feat: Enable proxy environment configuration
The 'deployment-environment' key in 'config.yml' can be used to define
environment variables (as key: values) to be set during deployment. This
is implemented to enable http/https proxy configuration but could be
used for anything that utilized environment variables.

The 'deployment-environment' dictionary is copied into
'playbooks/group_vars/all' as 'deployment_environment' (note the "-" is
changed to "_").

Change-Id: I4fb52828a409953edf886a2fc22649c1722d3128