fix: Ignore unneeded DHCPACK/DHCPRELEASE entries
lxc network constructs can generate DHCPACK/DHCPRELEASE entries not
related to the pxe network that we are trying to detect. Some of these
entries do not come with a mac or ip address.

watch_log.py now checks if DHCPACK and DHCPRELEASE entries have valid
mac and ip addresses before trying to return.

Change-Id: I932da82756a8863d00b6afd49f84903f983f48b6