feat: Verify sudo password before starting tasks
When using Ansible's "--ask-become-pass" (or its alias "-K") flags the
password is not validated until it is used. If a user types the password
incorrectly Cluster Genesis may not fail until after quite a few tasks
have completed. This commit adds an initial task that will only pass if
the sudo password is correct.

Change-Id: I9178c3fbffafc9d81fe0c78a0479a7033afb1f14