refactor: Add a passive mode message and prompt
When running in passive switch mode (using 'gen deploy-passive' and/or
'gen post-deploy-passive') an additional message and prompt is added to
inform the user that the pxe/data interfaces are about to be activated.
Once the interfaces are activated the user will need to coordinate
collecting the switch MAC address table within a (sometimes short)
window of opportunity.

Change-Id: I3184058773e120c105fd7f0954cbd3e0a8c16ac6