fix: Count BMC connections before issuing power on
In passive mode BMC IPMI connections are initially discovered by
attempting to send commands to all IPs with DHCP leases. Credentials are
tried in a "brute force" manner until a working set is found. This
commit verifies that the number of connections found is at least the
number of systems defined in 'config.yml'. Also, retry mechanisms are
added to the playbook to allow for "slow" connections and user
intervention before forcing a fail due to an insufficient number of IPMI
connections.

Change-Id: Iddd128864d70904f297e1ca6704d826de4798a49