fix: MLAG support for data switches
Adjust add_data_switch_port_macs parameters to accept a switch index
value. With the new switch class definitions, the iteration over each
switch in a rack is done in set_data_switch_config.py and not in
inventory.py. However, add_data_switch_port_macs relies on the index
value to determine if an adapter(eth11, eth13) belongs as part of an
mlag pair.

In addition, activate_client_interfaces.yml and
restore_client_interfaces.yml are updated to use ip link set, as opposed
to ifup/ifdown. This has proven to be more reliable when gathering MAC
addresses from the data switches.

Change-Id: I6002c31089cb0dbe0386efdccbf49d6cb62a9ef1