fix: Clear deployment environment during setup.yml
The 'deployment-environment' key in 'config.yml' is optional. When set
its value is copied into 'group_vars/all'. The only way to change or
clear the 'group_vars/all' variable for subsequent deploys was to
manually edit the file. This commit adds logic to clear the
'group_vars/all' variable and set it to the current value in
'config.yml'. If empty or undefined in 'config.yml' an empty value is
set in 'group_vars/all'. This ensures Ansible will be using the latest
'deployment-environment' defined by the user.

Change-Id: I91ac4e3486fdfa94c967422e00e34e97f0e20829