fix: Remove broken link to inventory file (#3)
The 'inventory.yml' file is kept inside of the LXC container. For
convenience a soft link is created on the deployer pointing to the
'inventory.yml' with the container file system. If the container is
destroyed the link will be broken. This patch will remove the link if it
is broken allowed it to be replaced by an empty file.

Change-Id: Ic1795c68738b137ec676ea1e5016e61862b5d375