fix: Create empty log file to ensure user is owner (#78)
If 'teardown' is run before a log file ('logs/gen') is created then the
log file will be created with 'root' user as the owner and without write
privileges for other users. Subsequent attempts to write to the log file
as a non-root user will fail. The 'install.sh' script will create 'logs/gen'
if it doesn't exist.