refactor: Improve software inventory prompts (#168)
Change the order and wording of the software inventory configuration
prompts so the "good" path to continue is the first option.

If a user has already completed the software inventory setup routine in
'--init-clients' then '--install' only needs to run validation. If
validation passes installation can continue without prompting, else if
validation fails the user will be presented with the software inventory
setup routine.