fix: Limit host list in 'lookup_interface_names' (#169)
The playbook 'lookup_interface_names.yml' is used to merge dictionaries
mapping all client node MACs to interface names.  It's possible for
environmental variables on localhost to cause jinja template errors when
they container "{{" "}}" characters. The play was run on all hosts but
only needs to be run on client nodes so this fix limits iteration over
client nodes only.