feat: Facilitate Anaconda license acceptance (#180)
Users are required to interactively accept the Anaconda license
agreement on at least one server per cluster. A single server is
selected (no significance as to which server) to run the manual
acceptance script. If the user accepts the license agreement on that
server then the acceptance script will be run quietly on the remaining
servers.

The IBM Knowledge Center instructions specify to install Anaconda into
the '/opt/anaconda2' directory.