fix: Add '.condarc' with installer channel (#198)
Populate a global '~/.condarc' channels list pointing to the installer
server.

During anaconda installation, if the Anaconda install directory exists 
(and has a bin/ dir) then skip the Anaconda installation script rather
than trying to update it. If there is no directory (with a bin/) then
Anaconda will install with a '-f' flag to overwrite any existing files
in the install directory.