feat: Allow client sudo password to be cached (#200)
The client sudo password (must be the same for all client nodes) is
encrypted (using Ansible Vault) and stored in 'software-vars.yml'. The 
sudo password itself is used as the vault encryption password. This
allows the user to provide the client sudo password only once when 'pup
software --install <module>' is invoked.