fix: Always unlock vault before running plays (#213)
Any 'pup' process instantiating a 'software' object will delete the
'.vault' file upon termination. If more than one 'pup' process is run
concurrently each process needs to verify the existence of the '.vault'
file before attempting to use it. The vault password is stored as an
attribute and thus can be used to re-unlock without re-prompting the
user.