fix: Add host key in get dependent package script
When using 'sshpass' to make 'ssh' calls, if host key is not present in
a user's 'known_hosts' file the script will fail without any output to
stdout or stderr (rc=6). If this situation is encountered within
'get-dependent-packages.sh', the user is left without any information on
why the script failed in a situation that is easily remedied. Logic is
now preset to check if the host key exists, if it does not then it is
added automatically (a message is printed to stdout to notify the user).