fix: Install PowerAI Enterprise license (#242)
* fix: Install PowerAI Enterprise license

The PowerAI Enterprise license is required to scale DDL beyond four
nodes.

The 'accept-powerai-enterprise-license.sh' script does not allow running
silently. There is a 'SILENT_MODE' variable in the script which is
assigned to '$1', which seems to indicate an intention to allow running
silently by passing an argument. The script also has logic to exit and
display usage if _any_ argument is passed, which prevents use of the
'SILENT_MODE'. To enable silent install the script must be modified to
set 'SILENT_MODE=yes'.

* fix: Run interactive license script on all hosts

The PowerAI Enterprise license acceptance is required to run
interactively on all hosts in the cluster.