fix: Configure inventory mount with SELinux label
If SELinux is running the 'inventory.yml' volume mounted in the
container needs a special 'Z' label (instead of 'rw').

Note: This mode has not been tested on a system _not_ running SELinux.
Logic may need to be added to switch this value from 'rw' to 'Z'
depending on the environment.