fix: Use 'dhclient' to build post-deploy mac table (#310)
* fix: Use 'dhclient' to build post-deploy mac table

Note: Ansible is updated to version 2.7.1. Re-running
      'scripts/install.sh' is *required* on deployers
      with an existing Power-Up install.

Assigning static IP addresses to client node data interfaces was not
generating MAC address table entries. Using dhclient to force a single
DHCP request (with a timeout of 1 second) works reliably.

If post-deploy '--gather-mac-addr' (gathers client node data interface
MAC addresses) fails to collect all addresses the user needs to both be
informed and given a recovery path.

If _any_ data interface MAC address is missing a table of _all_ data
interfaces is printed with the following columns: 'Switch', 'Port',
'Device', 'MAC'.

Once the table is printed the user is prompted to retry, continue (with
the missing MAC addresses), or exit the program.