refactor: Re-use Docker 'power-up' image (#384)
Docker 'power-up' images are tagged using a sha1sum of the 'Dockerfile'.
If a Docker image has already been built with the current 'Dockerfile'
(checksum of the file contents, _not_ git commits) then the image can be
re-used to save considerable time. As a result, scripts and python
virtual environment need to be updated during container creation
(instead of during image build).