fix: Prevent P9 package sync from P8 installer (#432)
When setting up the dependencies repository the user selects to sync
either P8 or P9 packages. If and _only if_ the installer is running the
same model processor the packages can be synced from locally enabled yum
repositories. This commit removes the option to sync when the installer
and client nodes do not have the same model of POWER processors.