fix: Check if any client needs restarting
When running pup from one of the client nodes ("self install") reboots
will be disabled on that node. A user could also manually disable
reboots on any node through the inventory 'pup_reboot' variable. If a
client required a kernel update it should be restarted before attempting
to install CUDA drivers. This commit checks 'needs-restarting' on every
client node (regardless of the 'pup_reboot' value) and prompts the user
to perform manual reboot if any client reports it necessary. The user
may select to continue with the install without performing any reboots
(the reboot _could_ be required for non-kernel/cuda related updates).