fix: Do not skip Ngnix 'status_prep()' call
A method call inside a logical expression should be placed _before_ the
'and' to ensure it is always evaluated. For example, Python will
evaluate the expression 'False and self.status_prep()' to 'False'
without calling the 'status_prep()' method. The expression
'self.status_prep() and False' also evaluates to 'False' _but_ the
'status_prep()' method is called.