feat: Improve final tasks run during self-install
When installing from a cluster node, "self-install", a manual reboot is
required before completing the installation. This commit creates a
'software-vars.yml' flag to allow re-running 'pup software wmla121
--install' to finish the install (user will be prompted to finish the
install, or if they don't want to finish they can clear the flag and/or
start the install from the beginning again).